refactor e2e summary
